Split off a number of entity files into seperate headers
