+++ src/openrct2/actions/BalloonPressAction.cpp
@@ -9,7 +9,8 @@
 
 #include "BalloonPressAction.h"
 
-#include "../world/Sprite.h"
+#include "../world/Balloon.h"
+#include "../world/Entity.h"
 #include "GameAction.h"
 
 BalloonPressAction::BalloonPressAction(uint16_t spriteIndex)
