+++ src/openrct2/libopenrct2.vcxproj
@@ -436,20 +436,25 @@
     <ClInclude Include="Version.h" />
     <ClInclude Include="windows\Intent.h" />
     <ClInclude Include="windows\tile_inspector.h" />
+    <ClInclude Include="world\Balloon.h" />
     <ClInclude Include="world\Banner.h" />
     <ClInclude Include="world\Climate.h" />
+    <ClInclude Include="world\Duck.h" />
     <ClInclude Include="world\Entity.h" />
     <ClInclude Include="world\EntityList.h" />
     <ClInclude Include="world\Entrance.h" />
     <ClInclude Include="world\Footpath.h" />
     <ClInclude Include="world\Fountain.h" />
     <ClInclude Include="world\LargeScenery.h" />
+    <ClInclude Include="world\Litter.h" />
     <ClInclude Include="world\Location.hpp" />
     <ClInclude Include="world\Map.h" />
     <ClInclude Include="world\MapAnimation.h" />
     <ClInclude Include="world\MapGen.h" />
     <ClInclude Include="world\MapHelpers.h" />
+    <ClInclude Include="world\MoneyEffect.h" />
     <ClInclude Include="world\Park.h" />
+    <ClInclude Include="world\Particle.h" />
     <ClInclude Include="world\Scenery.h" />
     <ClInclude Include="world\ScenerySelection.h" />
     <ClInclude Include="world\SmallScenery.h" />
