+++ src/openrct2/world/MoneyEffect.cpp
@@ -6,6 +6,7 @@
  *
  * OpenRCT2 is licensed under the GNU General Public License version 3.
  *****************************************************************************/
+#include "MoneyEffect.h"
 
 #include "../OpenRCT2.h"
 #include "../drawing/Drawing.h"
