+++ src/openrct2/world/Park.cpp
@@ -36,8 +36,8 @@
 #include "../util/Util.h"
 #include "../windows/Intent.h"
 #include "Entrance.h"
+#include "Litter.h"
 #include "Map.h"
-#include "Sprite.h"
 #include "Surface.h"
 
 #include <algorithm>
