+++ src/openrct2/world/Sprite.h
@@ -13,96 +13,16 @@
 #include "../common.h"
 #include "../peep/Peep.h"
 #include "../ride/Vehicle.h"
+#include "Balloon.h"
+#include "Duck.h"
 #include "Entity.h"
 #include "Fountain.h"
+#include "Litter.h"
+#include "MoneyEffect.h"
 #include "SpriteBase.h"
+#include "Particle.h"
 
-enum LitterType : uint8_t;
 class DataSerialiser;
-
-struct Litter : SpriteBase
-{
-    static constexpr auto cEntityType = EntityType::Litter;
-    LitterType SubType;
-    uint32_t creationTick;
-    void Serialise(DataSerialiser& stream);
-};
-
-struct Balloon : MiscEntity
-{
-    static constexpr auto cEntityType = EntityType::Balloon;
-    uint16_t popped;
-    uint8_t time_to_move;
-    uint8_t colour;
-
-    void Update();
-    void Pop();
-    void Press();
-    void Serialise(DataSerialiser& stream);
-};
-
-struct Duck : MiscEntity
-{
-    static constexpr auto cEntityType = EntityType::Duck;
-    enum class DuckState : uint8_t
-    {
-        FlyToWater,
-        Swim,
-        Drink,
-        DoubleDrink,
-        FlyAway,
-    };
-    int16_t target_x;
-    int16_t target_y;
-    DuckState state;
-
-    void Update();
-    uint32_t GetFrameImage(int32_t direction) const;
-    bool IsFlying();
-    void Remove();
-    void Serialise(DataSerialiser& stream);
-
-private:
-    void UpdateFlyToWater();
-    void UpdateSwim();
-    void UpdateDrink();
-    void UpdateDoubleDrink();
-    void UpdateFlyAway();
-};
-
-struct MoneyEffect : MiscEntity
-{
-    static constexpr auto cEntityType = EntityType::MoneyEffect;
-    uint16_t MoveDelay;
-    uint8_t NumMovements;
-    uint8_t Vertical;
-    money32 Value;
-    int16_t OffsetX;
-    uint16_t Wiggle;
-
-    static void CreateAt(money32 value, const CoordsXYZ& effectPos, bool vertical);
-    static void Create(money32 value, const CoordsXYZ& loc);
-    void Update();
-    std::pair<rct_string_id, money32> GetStringId() const;
-    void Serialise(DataSerialiser& stream);
-};
-
-struct VehicleCrashParticle : MiscEntity
-{
-    static constexpr auto cEntityType = EntityType::CrashedVehicleParticle;
-    uint16_t time_to_live;
-    uint8_t colour[2];
-    uint16_t crashed_sprite_base;
-    int16_t velocity_x;
-    int16_t velocity_y;
-    int16_t velocity_z;
-    int32_t acceleration_x;
-    int32_t acceleration_y;
-    int32_t acceleration_z;
-
-    void Update();
-    void Serialise(DataSerialiser& stream);
-};
 
 struct ExplosionFlare : MiscEntity
 {
@@ -114,13 +34,6 @@ struct ExplosionFlare : MiscEntity
 struct ExplosionCloud : MiscEntity
 {
     static constexpr auto cEntityType = EntityType::ExplosionCloud;
-    void Update();
-    void Serialise(DataSerialiser& stream);
-};
-
-struct CrashSplashParticle : MiscEntity
-{
-    static constexpr auto cEntityType = EntityType::CrashSplash;
     void Update();
     void Serialise(DataSerialiser& stream);
 };
@@ -178,22 +91,6 @@ enum
     SPRITE_FLAGS_PEEP_FLASHING = 1 << 9, // Deprecated: Use sprite_set_flashing/sprite_get_flashing instead.
 };
 
-enum LitterType : uint8_t
-{
-    LITTER_TYPE_SICK,
-    LITTER_TYPE_SICK_ALT,
-    LITTER_TYPE_EMPTY_CAN,
-    LITTER_TYPE_RUBBISH,
-    LITTER_TYPE_EMPTY_BURGER_BOX,
-    LITTER_TYPE_EMPTY_CUP,
-    LITTER_TYPE_EMPTY_BOX,
-    LITTER_TYPE_EMPTY_BOTTLE,
-    LITTER_TYPE_EMPTY_BOWL_RED,
-    LITTER_TYPE_EMPTY_DRINK_CARTON,
-    LITTER_TYPE_EMPTY_JUICE_CUP,
-    LITTER_TYPE_EMPTY_BOWL_BLUE,
-};
-
 constexpr const uint32_t SPATIAL_INDEX_SIZE = (MAXIMUM_MAP_SIZE_TECHNICAL * MAXIMUM_MAP_SIZE_TECHNICAL) + 1;
 constexpr const uint32_t SPATIAL_INDEX_LOCATION_NULL = SPATIAL_INDEX_SIZE - 1;
 
@@ -218,30 +115,9 @@ void sprite_clear_all_unused();
 void sprite_misc_update_all();
 void sprite_set_coordinates(const CoordsXYZ& spritePos, SpriteBase* sprite);
 void sprite_remove(SpriteBase* sprite);
-void litter_create(const CoordsXYZD& litterPos, LitterType type);
-void litter_remove_at(const CoordsXYZ& litterPos);
 uint16_t remove_floating_sprites();
 void sprite_misc_explosion_cloud_create(const CoordsXYZ& cloudPos);
 void sprite_misc_explosion_flare_create(const CoordsXYZ& flarePos);
-
-///////////////////////////////////////////////////////////////
-// Balloon
-///////////////////////////////////////////////////////////////
-void create_balloon(const CoordsXYZ& balloonPos, int32_t colour, bool isPopped);
-void balloon_update(Balloon* balloon);
-
-///////////////////////////////////////////////////////////////
-// Duck
-///////////////////////////////////////////////////////////////
-void create_duck(const CoordsXY& pos);
-void duck_press(Duck* duck);
-void duck_remove_all();
-
-///////////////////////////////////////////////////////////////
-// Crash particles
-///////////////////////////////////////////////////////////////
-void crashed_vehicle_particle_create(rct_vehicle_colour colours, const CoordsXYZ& vehiclePos);
-void crash_splash_create(const CoordsXYZ& splashPos);
 
 rct_sprite_checksum sprite_checksum();
